Insert manga into SQL table

+++ src/mal_backup_core/manga.rs
@@ -1,23 +1,26 @@
+use crate::schema::manga as MangaTable;
+use diesel::Insertable;
 use serde::Deserialize;
 
-#[derive(Debug, Deserialize)]
+#[derive(Debug, Deserialize, Insertable)]
+#[table_name = "MangaTable"]
 pub struct Manga {
-    pub mal_id: u32,
+    pub mal_id: i32,
     title: String,
     #[serde(rename = "type")]
     manga_type: String,
-    pub reading_status: u8,
-    score: u8,
-    read_chapters: u32,
-    read_volumes: u32,
-    total_chapters: u32,
-    total_volumes: u32,
-    publishing_status: u8,
+    pub reading_status: i32,
+    score: i32,
+    read_chapters: i32,
+    read_volumes: i32,
+    total_chapters: i32,
+    total_volumes: i32,
+    publishing_status: i32,
     is_rereading: bool,
     tags: Option<String>,
     start_date: Option<String>,
     end_date: Option<String>,
     read_start_date: Option<String>,
     read_end_date: Option<String>,
-    days: Option<u32>,
+    days: Option<i32>,
 }
